Split top-bar parity gates by ownership
