+++ Tests/QuillCodeParityTests/ParityTopBarSurfaceGateTests.swift
@@ -13,13 +13,7 @@ final class ParityTopBarSurfaceGateTests: QuillCodeParityTestCase {
             "normalizedUniqueModelIDs"
         ].forEach { Self.assertSource(builderText, contains: $0) }
         Self.assertSource(topBarBuilderText, contains: "WorkspaceModelCatalogSurfaceBuilder(")
-        [
-            "WorkspaceModelCatalogSurfaceBuilder(",
-            "func modelCategories(selectedModelID:",
-            "func modelOption(",
-            "func favoriteModelIDs()",
-            "func recentModelIDs("
-        ].forEach { Self.assertSource(surfaceText, excludes: $0) }
+        assertWorkspaceSurfaceAvoidsModelCatalogOwnership(surfaceText)
     }
 
     func testWorkspaceSurfaceDelegatesTopBarSurfaceContracts() throws {
@@ -27,30 +21,10 @@ final class ParityTopBarSurfaceGateTests: QuillCodeParityTestCase {
         let topBarText = try Self.appSourceText(named: "QuillCodeTopBarSurface.swift")
         let searchFilterText = try Self.appSourceText(named: "ModelCategorySearchFilter.swift")
 
-        [
-            "public struct TopBarSurface",
-            "public struct ModelCategorySurface",
-            "public struct ModelMetadataRowSurface",
-            "public struct ModelOptionSurface",
-            "filteredModelCategories",
-            "ModelCategorySearchFilter.filter"
-        ].forEach { Self.assertSource(topBarText, contains: $0) }
-        [
-            "enum ModelCategorySearchFilter",
-            "static func filter(",
-            "normalizedTerms"
-        ].forEach { Self.assertSource(searchFilterText, contains: $0) }
-        [
-            "includesFavoriteTerm",
-            "metadataRows.map"
-        ].forEach { Self.assertSource(topBarText, excludes: $0) }
-        [
-            "public struct TopBarSurface",
-            "public struct ModelCategorySurface",
-            "public struct ModelMetadataRowSurface",
-            "public struct ModelOptionSurface",
-            "filteredModelCategories"
-        ].forEach { Self.assertSource(surfaceText, excludes: $0) }
+        assertTopBarSurfaceContracts(topBarText)
+        assertModelSearchFilterContracts(searchFilterText)
+        assertTopBarSurfaceAvoidsSearchPolicy(topBarText)
+        assertWorkspaceSurfaceAvoidsTopBarRecords(surfaceText)
     }
 
     func testWorkspaceSurfaceDelegatesTopBarSurfaceBuilding() throws {
@@ -69,28 +43,49 @@ final class ParityTopBarSurfaceGateTests: QuillCodeParityTestCase {
         ].forEach { Self.assertSource(surfaceText, excludes: $0) }
     }
 
-    func testModelPickerWorkspaceIntegrationCoverageStaysFocused() throws {
-        let broadSurfaceTests = try Self.appTestSourceText(named: "WorkspaceSurfaceTests.swift")
-        let modelPickerTests = try Self.appTestSourceText(
-            named: "WorkspaceModelPickerSurfaceIntegrationTests.swift"
-        )
-        let topBarTests = try Self.appTestSourceText(named: "QuillCodeTopBarSurfaceTests.swift")
-        let modelPickerCases = [
-            "testSurfaceGroupsCustomModelCatalogByCategory",
-            "testTopBarFiltersModelCatalogByProviderCategoryAndModel",
-            "testSurfaceKeepsUnknownSelectedModelVisible",
-            "testModelPickerShowsRecentModelsAndBadges",
-            "testModelPickerShowsFavoriteModelsBeforeRecent"
-        ]
+    private func assertWorkspaceSurfaceAvoidsModelCatalogOwnership(_ source: String) {
+        [
+            "WorkspaceModelCatalogSurfaceBuilder(",
+            "func modelCategories(selectedModelID:",
+            "func modelOption(",
+            "func favoriteModelIDs()",
+            "func recentModelIDs("
+        ].forEach { Self.assertSource(source, excludes: $0) }
+    }
 
-        for testCase in modelPickerCases {
-            Self.assertSource(modelPickerTests, contains: "func \(testCase)")
-            Self.assertSource(broadSurfaceTests, excludes: "func \(testCase)")
-        }
-        Self.assertSource(topBarTests, contains: "func testModelOptionDecodesOlderPayloadWithoutBadges")
-        Self.assertSource(
-            broadSurfaceTests,
-            excludes: "func testModelOptionDecodesOlderPayloadWithoutBadges"
-        )
+    private func assertTopBarSurfaceContracts(_ source: String) {
+        [
+            "public struct TopBarSurface",
+            "public struct ModelCategorySurface",
+            "public struct ModelMetadataRowSurface",
+            "public struct ModelOptionSurface",
+            "filteredModelCategories",
+            "ModelCategorySearchFilter.filter"
+        ].forEach { Self.assertSource(source, contains: $0) }
+    }
+
+    private func assertModelSearchFilterContracts(_ source: String) {
+        [
+            "enum ModelCategorySearchFilter",
+            "static func filter(",
+            "normalizedTerms"
+        ].forEach { Self.assertSource(source, contains: $0) }
+    }
+
+    private func assertTopBarSurfaceAvoidsSearchPolicy(_ source: String) {
+        [
+            "includesFavoriteTerm",
+            "metadataRows.map"
+        ].forEach { Self.assertSource(source, excludes: $0) }
+    }
+
+    private func assertWorkspaceSurfaceAvoidsTopBarRecords(_ source: String) {
+        [
+            "public struct TopBarSurface",
+            "public struct ModelCategorySurface",
+            "public struct ModelMetadataRowSurface",
+            "public struct ModelOptionSurface",
+            "filteredModelCategories"
+        ].forEach { Self.assertSource(source, excludes: $0) }
     }
 }
