+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -22,7 +22,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `test:QuillCodeAppTests` | 203 | 31186 | 98 | A+ | `QuillCodeNativeHitTargetAuditTests.swift` A-<br>`WorkspaceRemoteProjectToolExecutorTests.swift` A-<br>`WorkspaceBrowserIntegrationTests.swift` A- |
 | `test:QuillCodeCoreTests` | 2 | 489 | 98 | A+ | `CoreModelTests.swift` A+<br>`ModelTokenUsageTests.swift` A+ |
 | `test:QuillCodeDesktopTests` | 3 | 1108 | 97 | A+ | `QuillCodeDesktopControllerSmokeTests.swift` A<br>`QuillCodeApprovalNotificationTests.swift` A+<br>`QuillCodeDesktopRenderedSmokeTests.swift` A+ |
-| `test:QuillCodeParityTests` | 94 | 11923 | 96 | A+ | `ParityWorkspaceExecutionToolGateTests.swift` B+<br>`ParityWorkspaceSurfaceGateTests.swift` B+<br>`ParityWorkspaceIntegrationGateTests.swift` B+ |
+| `test:QuillCodeParityTests` | 96 | 11959 | 96 | A+ | `ParityWorkspaceExecutionToolGateTests.swift` B+<br>`ParityWorkspaceSurfaceGateTests.swift` B+<br>`ParityWorkspaceIntegrationGateTests.swift` B+ |
 | `test:QuillCodePersistenceTests` | 1 | 301 | 91 | A- | `PersistenceTests.swift` A- |
 | `test:QuillCodeSafetyTests` | 5 | 1246 | 95 | A | `SafetyGitPullRequestPolicyTests.swift` A-<br>`SafetyGeneralPolicyTests.swift` A-<br>`SafetyDownloadPolicyTests.swift` A- |
 | `test:QuillCodeToolsTests` | 21 | 3474 | 98 | A+ | `GitHubPullRequestToolExecutorTests.swift` A-<br>`MCPStdioProberTests.swift` A-<br>`PTYProcessSessionTests.swift` A- |
@@ -962,7 +962,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 99 | 63 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityDesktopTaskCoordinationGateTests.swift` | 1 lines >120 chars |
 | A+ | 100 | 33 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityDesktopTrustedRouterAuthGateTests.swift` | no automated issues |
 | A+ | 100 | 51 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityFileMentionDirectoryGateTests.swift` | no automated issues |
-| A+ | 100 | 268 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityFocusedSuiteManifest.swift` | no automated issues |
+| A+ | 100 | 272 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityFocusedSuiteManifest.swift` | no automated issues |
 | A- | 91 | 94 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityGateTests.swift` | 3 lines >120 chars; 2 unsafe marker |
 | A+ | 98 | 343 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityGitPullRequestToolGateTests.swift` | 11 lines >120 chars |
 | A+ | 100 | 40 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityHTMLGateTests.swift` | no automated issues |
@@ -980,15 +980,17 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A- | 89 | 89 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityMCPGateTests.swift` | 60 lines >120 chars; 3 lines >160 chars |
 | A+ | 99 | 110 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityMergeTrainGateTests.swift` | 2 lines >120 chars |
 | A+ | 96 | 73 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityModelGateTests.swift` | 19 lines >120 chars; 1 lines >160 chars |
+| A+ | 100 | 30 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityModelPickerIntegrationGateTests.swift` | no automated issues |
 | A+ | 100 | 68 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeCompactHitTargetGateTests.swift` | no automated issues |
 | A- | 90 | 486 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeInteractionContractGateTests.swift` | 27 lines >120 chars; 11 lines >160 chars |
+| A+ | 100 | 27 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeModelPickerGateTests.swift` | no automated issues |
 | A+ | 100 | 67 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativePrimaryChromeHitTargetGateTests.swift` | no automated issues |
 | A+ | 100 | 130 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionAuditGateTests.swift` | no automated issues |
 | A+ | 100 | 60 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionAuditTestSupport.swift` | no automated issues |
 | A+ | 100 | 188 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionControlTargetTests.swift` | no automated issues |
 | A+ | 100 | 128 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionIconGestureTests.swift` | no automated issues |
 | A+ | 100 | 139 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionSpacingTests.swift` | no automated issues |
-| A+ | 100 | 68 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeTopBarChromeGateTests.swift` | no automated issues |
+| A+ | 100 | 48 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeTopBarChromeGateTests.swift` | no automated issues |
 | A+ | 99 | 67 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPackagedMacOSSmokeGateTests.swift` | 1 lines >120 chars |
 | A+ | 100 | 38 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightSettingsRuntimeGateTests.swift` | no automated issues |
 | A+ | 97 | 92 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityRemoteProjectToolGateTests.swift` | 16 lines >120 chars |
@@ -1006,7 +1008,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 218 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTestSupport.swift` | no automated issues |
 | A+ | 100 | 65 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityToolGateTests.swift` | no automated issues |
 | A+ | 100 | 91 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTopBarPresentationGateTests.swift` | no automated issues |
-| A+ | 100 | 96 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTopBarSurfaceGateTests.swift` | no automated issues |
+| A+ | 100 | 91 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTopBarSurfaceGateTests.swift` | no automated issues |
 | B+ | 87 | 129 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTrustedRouterGateTests.swift` | 52 lines >120 chars; 6 lines >160 chars |
 | B+ | 86 | 131 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceCommandGateTests.swift` | 67 lines >120 chars; 17 lines >160 chars |
 | A+ | 100 | 82 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceConfigurationGateTests.swift` | no automated issues |
