+++ docs/DECISIONS.md
@@ -542,3 +542,4 @@
 - Project parity gates are split by ownership. Project metadata, WorkspaceModel project API, pure loader ownership, and instruction-scope contracts stay in `ParityWorkspaceProjectGateTests`; local/remote/PR integration-suite placement lives in `ParityWorkspaceProjectIntegrationGateTests`; worktree integration placement and worktree source contracts live in `ParityWorkspaceWorktreeGateTests`. The focused-suite manifest must include every test in those files, including project API and instruction-scope gates that were previously omitted.
 - Workspace model thread parity gates are split by behavior family. Project-context refresh, seed/summary construction, creation records, lifecycle mutation/persistence, and configuration transitions each live in separate parity files. Keep new thread/configuration source-contract checks with the family they protect instead of re-growing a broad model/thread catch-all.
 - HTML renderer delegation parity gates are split by renderer family. Tool-card, top-bar, terminal, secondary-pane, transcript/review, and sidebar HTML contracts each live in focused parity files, while `ParityHTMLGateTests` remains only the broad-suite sentinel. Add new HTML source-contract checks beside the renderer family they protect.
+- Top-bar parity gates are split by presentation/status, native chrome, surface/model catalog, native model picker, and integration ownership. The focused-suite manifest must list every top-bar/model-picker gate; do not use a broad top-bar catch-all for unrelated status, UI chrome, DTO, and integration contracts.
